unspecified-css-subresources/list-style-image.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>list-style blocks load</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<ul id=foo><li>XXX</ul>
<style>
#foo {
  list-style: url("/images/blue.png?pipe=trickle(d5)");
}
</style>
<script>
var t1 = performance.timing.responseStart;

async_test(function(t) {
window.addEventListener("load", t.step_func_done(function() {
    var t2 = Date.now();
    var diff = t2 - t1;
    assert_greater_than(diff, 5000, "ms after style");
  }));
}, "load waits for list-style");
</script>
